Backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
sqlalchemy
pandas
psycopg2-binary
ortools